Поправил интерфейс i-user-card-data
