--- sandbox/projects/h21-be-ui-kit/src/dto/i-user-card-data.ts
@@ -1,18 +1,0 @@
-export interface IUserCardAction {
-	name: string;
-	label: string;
-	icon: string;
-	route: string;
-	type: 'link' | 'button';
-}
-
-export interface IUserCardUser {
-	name: string;
-	email: string;
-	avatarUrl: string;
-}
-
-export interface IUserCardData{
-	user: IUserCardUser;
-	actions: IUserCardAction[];
-}
