+++ sandbox/projects/h21-be-ui-kit/src/lib/h21-user-card/dto/i-user-card-action.ts
@@ -1,0 +1,7 @@
+export interface IUserCardAction {
+	name: string;
+	label: string;
+	icon: string;
+	route: string;
+	type: 'link' | 'button';
+}
