+++ sandbox/projects/h21-be-ui-kit/src/lib/h21-user-card/dto/i-user-card-data.ts
@@ -1,0 +1,7 @@
+import {IUserCardAction} from "./i-user-card-action";
+import {IUserCardUser} from "./i-user-card-user";
+
+export interface IUserCardData{
+	user: IUserCardUser;
+	actions: IUserCardAction[];
+}
