+++ sandbox/projects/h21-be-ui-kit/src/lib/h21-user-card/dto/i-user-card-user.ts
@@ -1,0 +1,5 @@
+export interface IUserCardUser {
+	name: string;
+	email: string;
+	avatarUrl: string;
+}
